internal/processor/processor_test.go
package processor_test

import (
	"testing"

	"github.com/rithindattag/realtime-streaming-api/internal/processor"
	"github.com/rithindattag/realtime-streaming-api/pkg/logger"
	"github.com/stretchr/testify/assert"
)

// TestProcessMessage tests the ProcessMessage function
func TestProcessMessage(t *testing.T) {
	// Initialize logger and processor
	log := logger.NewLogger()
	p := processor.NewProcessor(log)

	// Test cases
	testCases := []struct {
		name     string
		input    []byte
		expected []byte
	}{
		// Add your test cases here
	}

	// Run test cases
	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			result := p.ProcessMessage(tc.input)
			assert.Equal(t, tc.expected, result)
		})
	}
}
